Inicializando modulo de persistencia: Adicionando ao pom xml a dependencia do mysql. Adicionando atributo id nas entidades do Alugame. Criando classe de Banco com metodo de conexao. Criando classe JogoDAO com metodos de busca e insert.

+++ src/main/kotlin/br/com/alura/alugames/modelo/IRecomenda.kt
@@ -3,6 +3,6 @@ package br.com.alura.alugames.modelo
 import java.math.BigDecimal
 
 interface IRecomenda {
-    val media: BigDecimal
+    val media: Double
 
     fun recomendar(nota: Int)}
